requirements.txt
asgiref==3.2.3
Django==3.0.4
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
pytz==2019.3
sqlparse==0.3.1
